feat: add optional auth to view event (#17)
optional encrypted password field stored for the event.

When getting the event to view, if password is set it adds a basic auth middleware with username `guest` and the password set by the owner of the event.

When generating QR code for an event, if it has a password - optionally encode this into the QR so that guests can scan the QR code and open straight into the event

fixes #16

Reviewed-on: https://gitea.tempel-koi.ts.net/jj-style/eventpix/pulls/17
Co-authored-by: JJ Style <style.jj@pm.me>
Co-committed-by: JJ Style <style.jj@pm.me>